Fix 32/64 bit MP message structure packing issues
- 32 bit and 64 bit should work the same way now (need to test).
  The server compiles 32 bit at the moment.
- 8 bytes from the NetMessage header were removed, and the protocol
  version was bumped.

+++ mpshared/netmessage.cpp
@@ -7,7 +7,7 @@
 
 namespace wi {
 
-long gatGameSpeeds[15] = {
+int gatGameSpeeds[15] = {
     64, 48, 32, 24, 16, 12, 10, 8, 7, 6, 5, 4, 3, 2, 1
 };
